+++ lib/main.dart
@@ -1,5 +1,4 @@
-import 'package:fl_chart/fl_chart.dart';
-import 'package:flutter/gestures.dart';
+import 'package:coneg/ui/dashboard.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -14,14 +13,92 @@ class Home extends StatefulWidget {
 }
 
 class _HomeState extends State<Home> {
-  int touchedIndex, doubleClick;
-  double heightCont = 0;
-  Color colorCont;
+  final PageController pageController;
+  _HomeState(this.pageController);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+        drawer: Drawer(
+          child: ListView(
+            padding: EdgeInsets.zero,
+            children: <Widget>[
+              DrawerHeader(
+                decoration: BoxDecoration(
+                  color: Color(0xFF23A39B),
+                ),
+                child: Container(
+                  margin: EdgeInsets.all(30),
+                  alignment: Alignment.center,
+                  child: Text(
+                    'ConEg - Menu',
+                    style: TextStyle(
+                        fontSize: 30,
+                        fontWeight: FontWeight.bold,
+                        color: Colors.white),
+                  ),
+                  decoration: BoxDecoration(
+                    color: Color(0xFF1F41B4),
+                    border: Border.all(
+                      color: Color(0xFF970062),
+                      width: 4.0,
+                    ),
+                    borderRadius: BorderRadius.circular(15),
+                    boxShadow: [
+                      BoxShadow(
+                        color: Color(0xFF970062).withOpacity(0.8),
+                        spreadRadius: 5,
+                        blurRadius: 7, // changes position of shadow
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+              ListTile(
+                title: Text('Dashboard'),
+                leading: Icon(Icons.dashboard_rounded),
+                onTap: () {
+                  Navigator.of(context).push(
+                      context, MaterialPageRoute(builder: () => Dashboard()));
+                },
+              ),
+              ListTile(
+                title: Text('Cadastro de Pessoas'),
+                leading: Icon(Icons.add_box_rounded),
+                onTap: () {
+                  // Update the state of the app.
+                  // ...
+                },
+              ),
+              ListTile(
+                title: Text('Configuração de Notificação'),
+                leading: Icon(Icons.notification_important_rounded),
+                onTap: () {
+                  // Update the state of the app.
+                  // ...
+                },
+              ),
+              ListTile(
+                title: Text('Configuração do Administrador'),
+                leading: Icon(Icons.miscellaneous_services_rounded),
+                onTap: () {
+                  // Update the state of the app.
+                  // ...
+                },
+              ),
+              ListTile(
+                title: Text('Sair'),
+                leading: Icon(Icons.exit_to_app_rounded),
+                onTap: () {
+                  // Update the state of the app.
+                  // ...
+                },
+              ),
+            ],
+          ),
+        ),
         appBar: AppBar(
+          backgroundColor: Color(0xFF23A39B),
           title: Padding(
             padding: EdgeInsets.all(10),
             child: Image.asset(
@@ -36,134 +113,11 @@ class _HomeState extends State<Home> {
         body: Column(
           crossAxisAlignment: CrossAxisAlignment.stretch,
           children: <Widget>[
-            Padding(
-              padding: EdgeInsets.all(20),
-              child: Container(
-                height: 500,
-                width: 500,
-                child: PieChart(
-                  PieChartData(
-                      pieTouchData: //TODO - Manipulate Variables
-                          PieTouchData(touchCallback: (pieTouchResponse) {
-                        setState(() {
-                          //Check clicks
-                          final desiredTouch = pieTouchResponse.touchInput
-                                  is! PointerExitEvent &&
-                              pieTouchResponse.touchInput is! PointerUpEvent;
-
-                          if (desiredTouch &&
-                              pieTouchResponse.touchedSection != null) {
-                            //Not clicks and section is not null
-                            touchedIndex = pieTouchResponse
-                                .touchedSection.touchedSectionIndex;
-                          } else if (!desiredTouch &&
-                              pieTouchResponse.touchedSection != null) {
-                            if (doubleClick ==
-                                pieTouchResponse
-                                    .touchedSection.touchedSectionIndex) {
-                              heightCont = 0;
-                              doubleClick = null;
-                            } else {
-                              doubleClick = pieTouchResponse
-                                  .touchedSection.touchedSectionIndex;
-                              //There's a click and section is not null
-                              colorCont = pieTouchResponse
-                                  .touchedSection.touchedSection.color
-                                  .withOpacity(0.6);
-                              heightCont = 500;
-                            }
-                          } else {
-                            touchedIndex = -1;
-                          }
-                        });
-                      }),
-                      startDegreeOffset: 180,
-                      sections: showingSections(),
-                      centerSpaceRadius: 0),
-                  swapAnimationDuration: Duration(milliseconds: 500),
-                  swapAnimationCurve: Curves.linear,
-                ),
-              ),
-            ),
-            Padding(
-              padding: EdgeInsets.all(20),
-              child: Container(
-                alignment: Alignment.center,
-                child: AnimatedContainer(
-                  duration: Duration(milliseconds: 500),
-                  height: heightCont,
-                  width: 800,
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(20),
-                    color: colorCont,
-                  ),
-                ),
-              ),
+            Container(
+              color: Color(0xFF17DFD3),
+              height: 50,
             ),
           ],
         ));
-  }
-
-  List<PieChartSectionData> showingSections() {
-    return List.generate(
-      4,
-      (i) {
-        final isTouched = i == touchedIndex;
-        final double opacity = isTouched ? 1 : 0.6;
-        switch (i) {
-          case 0:
-            return PieChartSectionData(
-              color: const Color(0xff0293ee).withOpacity(opacity),
-              value: 25,
-              title: 'Test',
-              radius: 200,
-              showTitle: true,
-              titleStyle: TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.bold,
-                  color: const Color(0xff044d7c)),
-              titlePositionPercentageOffset: 0.55,
-            );
-          case 1:
-            return PieChartSectionData(
-              color: const Color(0xfff8b250).withOpacity(opacity),
-              value: 25,
-              title: '',
-              radius: 200,
-              titleStyle: TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.bold,
-                  color: const Color(0xff90672d)),
-              titlePositionPercentageOffset: 0.55,
-            );
-          case 2:
-            return PieChartSectionData(
-              color: const Color(0xff845bef).withOpacity(opacity),
-              value: 25,
-              title: '',
-              radius: 200,
-              titleStyle: TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.bold,
-                  color: const Color(0xff4c3788)),
-              titlePositionPercentageOffset: 0.6,
-            );
-          case 3:
-            return PieChartSectionData(
-              color: const Color(0xff13d38e).withOpacity(opacity),
-              value: 25,
-              title: '',
-              radius: 200,
-              titleStyle: TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.bold,
-                  color: const Color(0xff0c7f55)),
-              titlePositionPercentageOffset: 0.55,
-            );
-          default:
-            return null;
-        }
-      },
-    );
   }
 }
